Try to support pseudo3d cone light
