+++ src/box2dLight/ConeLight.java
@@ -45,6 +45,11 @@ public class ConeLight extends PositionalLight {
 	
 	@Override
 	public void update () {
+		if (rayHandler.pseudo3d) {
+			prepareFixtureData();
+			updateDynamicShadowMeshes();
+		}
+
 		updateBody();
 		if (dirty) setEndPoints();
