+++ src/box2dLight/PointLight.java
@@ -70,189 +70,6 @@ public class PointLight extends PositionalLight {
 		updateMesh();
 	}
 
-	protected void updateDynamicShadowMeshes() {
-		int meshInd = 0;
-		float colBits = rayHandler.ambientLight.toFloatBits();
-		for (Fixture fixture : affectedFixtures) {
-			LightData data = (LightData)fixture.getUserData();
-			if (data == null || fixture.isSensor()) continue;
-
-			int size = 0;
-			float l = 0f;
-			float f = 1f / data.shadowsDropped;
-
-			float startColBits = rayHandler.shadowColorInterpolation ?
-					Color.BLACK.lerp(rayHandler.ambientLight, 1-f).toFloatBits() :
-					zeroColorBits;
-			float endColBits = rayHandler.shadowColorInterpolation ?
-					Color.WHITE.lerp(rayHandler.ambientLight, 1-f).toFloatBits() :
-					colBits;
-
-			Shape fixtureShape = fixture.getShape();
-			Shape.Type type = fixtureShape.getType();
-			Body body = fixture.getBody();
-			center.set(body.getWorldCenter());
-
-			if (type == Shape.Type.Polygon || type == Shape.Type.Chain) {
-				boolean isPolygon = (type == Shape.Type.Polygon);
-				ChainShape cShape = isPolygon ?
-						null : (ChainShape)fixtureShape;
-				PolygonShape pShape = isPolygon ?
-						(PolygonShape)fixtureShape : null;
-				int vertexCount = isPolygon ?
-						pShape.getVertexCount() : cShape.getVertexCount();
-				int minN = -1;
-				int maxN = -1;
-				int minDstN = -1;
-				float minDst = Float.POSITIVE_INFINITY;
-				boolean hasGasp = false;
-				tmpVerts.clear();
-				for (int n = 0; n < vertexCount; n++) {
-					if (isPolygon) {
-						pShape.getVertex(n, tmpVec);
-					} else {
-						cShape.getVertex(n, tmpVec);
-					}
-					tmpVec.set(body.getWorldPoint(tmpVec));
-					tmpVerts.add(tmpVec.cpy());
-					tmpEnd.set(tmpVec).sub(start).limit2(0.0001f).add(tmpVec);
-					if (fixture.testPoint(tmpEnd)) {
-						if (minN == -1) minN = n;
-						maxN = n;
-						hasGasp = true;
-						continue;
-					}
-
-					float currDist = tmpVec.dst2(start);
-					if (currDist < minDst) {
-						minDst = currDist;
-						minDstN = n;
-					}
-				}
-
-				ind.clear();
-				if (!hasGasp) {
-					tmpVec.set(tmpVerts.get(minDstN));
-					for (int n = minDstN; n < vertexCount; n++) {
-						ind.add(n);
-					}
-					for (int n = 0; n < minDstN; n++) {
-						ind.add(n);
-					}
-					if (Intersector.pointLineSide(start, center, tmpVec) > 0) {
-						ind.reverse();
-						ind.insert(0, ind.pop());
-					}
-				} else if (minN == 0 && maxN == vertexCount - 1) {
-					for (int n = maxN - 1; n > minN; n--) {
-						ind.add(n);
-					}
-				} else {
-					for (int n = minN - 1; n > -1; n--) {
-						ind.add(n);
-					}
-					for (int n = vertexCount - 1; n > maxN ; n--) {
-						ind.add(n);
-					}
-				}
-
-				for (int n : ind.toArray()) {
-					tmpVec.set(tmpVerts.get(n));
-
-					float dst = tmpVec.dst(start);
-					l = data.getLimit(dst, pseudo3dHeight, distance);
-					tmpEnd.set(tmpVec).sub(start).setLength(l).add(tmpVec);
-
-					segments[size++] = tmpVec.x;
-					segments[size++] = tmpVec.y;
-					segments[size++] = startColBits;
-					segments[size++] = f;
-
-					segments[size++] = tmpEnd.x;
-					segments[size++] = tmpEnd.y;
-					segments[size++] = endColBits;
-					segments[size++] = f;
-				}
-			} else if (type == Shape.Type.Circle) {
-				CircleShape shape = (CircleShape)fixtureShape;
-				float r = shape.getRadius();
-				float dst = tmpVec.set(center).dst(start);
-				float a = (float) Math.acos(r/dst);
-				l = data.getLimit(dst, pseudo3dHeight, distance);
-
-				tmpVec.set(start).sub(center).clamp(r, r).rotateRad(a);
-				tmpStart.set(center).add(tmpVec);
-
-				float angle = (MathUtils.PI2 - 2f * a) /
-						RayHandler.CIRCLE_APPROX_POINTS;
-				for (int k = 0; k < RayHandler.CIRCLE_APPROX_POINTS; k++) {
-					tmpStart.set(center).add(tmpVec);
-					segments[size++] = tmpStart.x;
-					segments[size++] = tmpStart.y;
-					segments[size++] = startColBits;
-					segments[size++] = f;
-
-					tmpEnd.set(tmpStart).sub(start).setLength(l).add(tmpStart);
-					segments[size++] = tmpEnd.x;
-					segments[size++] = tmpEnd.y;
-					segments[size++] = endColBits;
-					segments[size++] = f;
-
-					tmpVec.rotateRad(angle);
-				}
-			} else if (type == Shape.Type.Edge) {
-				EdgeShape shape = (EdgeShape)fixtureShape;
-
-				shape.getVertex1(tmpVec);
-				tmpVec.set(body.getWorldPoint(tmpVec));
-				float dst = tmpVec.dst(start);
-				l = data.getLimit(dst, pseudo3dHeight, distance);
-
-				segments[size++] = tmpVec.x;
-				segments[size++] = tmpVec.y;
-				segments[size++] = startColBits;
-				segments[size++] = f;
-
-				tmpEnd.set(tmpVec).sub(start).setLength(l).add(tmpVec);
-				segments[size++] = tmpEnd.x;
-				segments[size++] = tmpEnd.y;
-				segments[size++] = endColBits;
-				segments[size++] = f;
-
-				shape.getVertex2(tmpVec);
-				tmpVec.set(body.getWorldPoint(tmpVec));
-				dst = tmpVec.dst(start);
-				l = data.getLimit(dst, pseudo3dHeight, distance);
-
-				segments[size++] = tmpVec.x;
-				segments[size++] = tmpVec.y;
-				segments[size++] = startColBits;
-				segments[size++] = f;
-
-				tmpEnd.set(tmpVec).sub(start).setLength(l).add(tmpVec);
-				segments[size++] = tmpEnd.x;
-				segments[size++] = tmpEnd.y;
-				segments[size++] = endColBits;
-				segments[size++] = f;
-			}
-
-			Mesh mesh = null;
-			if (meshInd >= dynamicShadowMeshes.size) {
-				mesh = new Mesh(
-						Mesh.VertexDataType.VertexArray, false, RayHandler.MAX_SHADOW_VERTICES, 0,
-						new VertexAttribute(VertexAttributes.Usage.Position, 2, "vertex_positions"),
-						new VertexAttribute(VertexAttributes.Usage.ColorPacked, 4, "quad_colors"),
-						new VertexAttribute(VertexAttributes.Usage.Generic, 1, "s"));
-				dynamicShadowMeshes.add(mesh);
-			} else {
-				mesh = dynamicShadowMeshes.get(meshInd);
-			}
-			mesh.setVertices(segments, 0, size);
-			meshInd++;
-		}
-		dynamicShadowMeshes.truncate(meshInd);
-	}
-
 	/**
 	 * Sets light distance
 	 *
